requirements.txt
josn
chainlit
dotenv
google-generativeai
groq